kernel/sched: Use the highest capacity CPU when no CPU fits a task If a waking task doesn't fit on any CPU then it should be placed on the highest capacity CPU, otherwise the active load balancer will need to deal with moving the task to a higher capacity CPU later. This hurts performance both due to the suboptimal task placement and migration overhead incurred by active load balancing to fix the load imbalance later.
Selecting the least utilized CPU for heavy tasks is inherently flawed
because a higher capacity CPU might be overutilized while a lower capacity
CPU isn't, which can cause heavy tasks to be suboptimally placed on a lower
capacity CPU that cannot satisfy the heavy task's compute requirements.

Regardless of the highest capacity CPU's utilization ratio, if it needs
small tasks moved off of it to accommodate a heavy task, then the load
balancer can take care of it. So just make the right decision in feec()
from the get-go for the heavy task and wake it on the highest-capacity CPU.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>